Fix warnings

+++ src/tests/roc_lib/test_sender_receiver.cpp
@@ -137,7 +137,7 @@ TEST_GROUP(sender_receiver) {
         while(s_last == 0){
             size_t i = 0;
             ipacket++;
-            CHECK(roc_receiver_read(recv, rx_buff, packet_len) == packet_len);
+            CHECK(roc_receiver_read(recv, rx_buff, packet_len) == (ssize_t)packet_len);
             if (seek_first) {
                 for (; i < packet_len && fabs(double(rx_buff[i])) < 1e-9; i++, s_first++)
                     {}
@@ -151,13 +151,20 @@ TEST_GROUP(sender_receiver) {
                     if (inner_cntr >= len) {
                         CHECK(fabs(double(rx_buff[i])) < 1e-9);
                         s_last = inner_cntr + s_first;
-                        roc_log(LogInfo, "FINISH: s_first: %lu, s_last: %lu, inner_cntr: %lu",
-                            s_first, s_last, inner_cntr);
+                        roc_log(LogInfo,
+                            "FINISH: s_first: %lu, s_last: %lu, inner_cntr: %lu",
+                                (unsigned long)s_first,
+                                (unsigned long)s_last,
+                                (unsigned long)inner_cntr);
                         break;
                     } else if (fabs(double(original[inner_cntr] - rx_buff[i])) > 1e-9) {
                         char sbuff[256];
-                        int sbuff_i = snprintf(sbuff, sizeof(sbuff), "Failed comparing samples #%lu\n\npacket_num: %lu\n", inner_cntr, ipacket);
-                        snprintf(&sbuff[sbuff_i], sizeof(sbuff)-(size_t)sbuff_i, "original: %f,\treceived: %f\n", original[inner_cntr], rx_buff[i]);
+                        int sbuff_i = snprintf(sbuff, sizeof(sbuff),
+                            "Failed comparing samples #%lu\n\npacket_num: %lu\n",
+                                (unsigned long)inner_cntr, (unsigned long)ipacket);
+                        snprintf(&sbuff[sbuff_i], sizeof(sbuff)-(size_t)sbuff_i,
+                            "original: %f,\treceived: %f\n",
+                                (double)original[inner_cntr], (double)rx_buff[i]);
                         FAIL(sbuff);
                     } else {
                         inner_cntr++;
